api-extractor-be/requirements.txt
crewai
crewai-tools
python-dotenv
streamlit
FastAPI
firebase-admin
